requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Markdown==3.4.3
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
